+++ lib/common/point3.ts
@@ -1,0 +1,8 @@
+import { distance } from "@tscircuit/soup"
+import { z } from "zod"
+
+export const point3 = z.object({
+  x: distance,
+  y: distance,
+  z: distance,
+})
