--- lib/utils/direction.ts
@@ -1,3 +1,0 @@
-import { z } from "zod"
-export const direction = z.enum(["up", "down", "left", "right"])
-export type Direction = "up" | "down" | "left" | "right"
