+++ package.json
@@ -24,6 +24,7 @@
     "@types/react": "^18.3.2",
     "ava": "^6.1.3",
     "expect-type": "^0.20.0",
+    "madge": "^8.0.0",
     "react": "^18.3.1",
     "tsup": "^8.0.2",
     "tsx": "^4.10.2",
